+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 -   Improved layout of example code output refresh.
 -   Tidier spacing of project preview lists.
 -   [#468](https://github.com/wordplaydev/wordplay/issues/468) Fixed example dragging from tutorial dialog.
+-   [#465](https://github.com/wordplaydev/wordplay/issues/465) Fixed ARIA roles for output.
 
 ### Added